app/requirements/dev.txt
-r prod.txt
black
isort
pylint-django
faker==11.1.0